refactor: move shared ffi testing utils into module
